requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
black==22.3.0
certifi==2020.6.20
click==8.1.2
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
wincertstore==0.2
wrapt==1.14.0
